Norminal to nominal scale name
I did not read the reference, but the scale that applies to items only being i
related among each other per equality relation is called nominal and not norminal.

This is a nit only, but as documentation targets the readers understanding ir may be worth the change.

Nominal comes from the Latin nomen which means name, we can at least identify and thus name the
Items. No norm is possible, and no oredering is defined either - one has to apply a convention.